+++ bbangzip-api/build.gradle
@@ -10,16 +10,9 @@ dependencies {
     implementation project(":bbangzip-domain")
     implementation project(":bbangzip-external")
     implementation project(":bbangzip-common")
+    implementation project(":bbangzip-auth")
 
     implementation 'mysql:mysql-connector-java:8.0.33'
-
-    //Security
-    implementation 'org.springframework.boot:spring-boot-starter-security'
-
-    //JWT
-    implementation group: 'io.jsonwebtoken', name: 'jjwt-api', version: '0.11.5'
-    implementation group: 'io.jsonwebtoken', name: 'jjwt-impl', version: '0.11.5'
-    implementation group: 'io.jsonwebtoken', name: 'jjwt-jackson', version: '0.11.5'
 
     // Validation
     implementation 'org.springframework.boot:spring-boot-starter-validation'
