--- bbangzip-api/src/main/java/org/sopt/auth/exception/BbangzipBusinessException.java
@@ -1,16 +1,0 @@
-package org.sopt.auth.exception;
-
-import org.sopt.code.ErrorCode;
-import org.springframework.http.HttpStatus;
-
-public class BbangzipBusinessException extends BbangzipAuthException {
-    public BbangzipBusinessException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-
-    @Override
-    public HttpStatus getStatus() {
-        return HttpStatus.BAD_REQUEST;
-    }
-}
-
