--- bbangzip-api/src/main/java/org/sopt/auth/exception/UnAuthorizedException.java
@@ -1,16 +1,0 @@
-package org.sopt.auth.exception;
-
-import org.sopt.code.ErrorCode;
-import org.springframework.http.HttpStatus;
-
-public class UnAuthorizedException extends BbangzipAuthException {
-
-    public UnAuthorizedException(ErrorCode errorCode) {
-        super(errorCode);
-    }
-
-    @Override
-    public HttpStatus getStatus() {
-        return null;
-    }
-}
