--- bbangzip-api/src/main/java/org/sopt/auth/jwt/dto/ReissueJwtTokensDto.java
@@ -1,10 +1,0 @@
-package org.sopt.auth.jwt.dto;
-
-import lombok.Builder;
-
-@Builder
-public record ReissueJwtTokensDto(
-        String accessToken,
-        String refreshToken
-) {
-}
