--- bbangzip-api/src/main/java/org/sopt/auth/security/logsupport/ErrorLogger.java
@@ -1,17 +1,0 @@
-package org.sopt.auth.security.logsupport;
-
-import lombok.extern.slf4j.Slf4j;
-import org.sopt.code.ErrorCode;
-import org.springframework.stereotype.Component;
-
-@Slf4j
-@Component
-public class ErrorLogger {
-    public void log(ErrorCode errorCode, Exception e) {
-        if (errorCode.getHttpStatus().is4xxClientError()) {
-            log.warn("Client error [{}]: {}", errorCode.getCode(), e.getMessage());
-        } else if (errorCode.getHttpStatus().is5xxServerError()) {
-            log.error("Server error [{}]: {}", errorCode.getCode(), e.getMessage(), e);
-        }
-    }
-}
