+++ bbangzip-api/src/main/java/org/sopt/test/TestController.java
@@ -2,9 +2,9 @@ package org.sopt.test;
 
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
-import org.sopt.auth.annotation.UserId;
-import org.sopt.auth.jwt.JwtTokenProvider;
-import org.sopt.auth.jwt.dto.JwtTokensDto;
+import org.sopt.annotation.UserId;
+import org.sopt.jwt.JwtTokenProvider;
+import org.sopt.jwt.dto.JwtTokensDto;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
