+++ bbangzip-auth/build.gradle
@@ -1,0 +1,15 @@
+dependencies {
+    implementation project(":bbangzip-common")
+
+    //Security
+    implementation 'org.springframework.boot:spring-boot-starter-security'
+
+    //JWT
+    implementation group: 'io.jsonwebtoken', name: 'jjwt-api', version: '0.11.5'
+    implementation group: 'io.jsonwebtoken', name: 'jjwt-impl', version: '0.11.5'
+    implementation group: 'io.jsonwebtoken', name: 'jjwt-jackson', version: '0.11.5'
+
+}
+
+bootJar { enabled = false }
+jar { enabled = true }
