+++ bbangzip-auth/src/main/java/org/sopt/exception/BbangzipBusinessException.java
@@ -1,0 +1,16 @@
+package org.sopt.exception;
+
+import org.sopt.code.ErrorCode;
+import org.springframework.http.HttpStatus;
+
+public class BbangzipBusinessException extends BbangzipAuthException {
+    public BbangzipBusinessException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+
+    @Override
+    public HttpStatus getStatus() {
+        return HttpStatus.BAD_REQUEST;
+    }
+}
+
