+++ bbangzip-auth/src/main/java/org/sopt/exception/TokenNotFoundException.java
@@ -1,0 +1,18 @@
+package org.sopt.exception;
+
+import lombok.Getter;
+import org.sopt.code.ErrorCode;
+import org.springframework.http.HttpStatus;
+
+@Getter
+public class TokenNotFoundException extends BbangzipAuthException {
+
+    public TokenNotFoundException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+
+    @Override
+    public HttpStatus getStatus() {
+        return HttpStatus.NOT_FOUND;
+    }
+}
