+++ bbangzip-auth/src/main/java/org/sopt/exception/UnAuthorizedException.java
@@ -1,0 +1,16 @@
+package org.sopt.exception;
+
+import org.sopt.code.ErrorCode;
+import org.springframework.http.HttpStatus;
+
+public class UnAuthorizedException extends BbangzipAuthException {
+
+    public UnAuthorizedException(ErrorCode errorCode) {
+        super(errorCode);
+    }
+
+    @Override
+    public HttpStatus getStatus() {
+        return null;
+    }
+}
