+++ bbangzip-auth/src/main/java/org/sopt/jwt/authentication/UserAuthenticationFactory.java
@@ -1,0 +1,21 @@
+package org.sopt.jwt.authentication;
+
+import io.jsonwebtoken.Claims;
+import jakarta.servlet.http.HttpServletRequest;
+import org.sopt.exception.AuthConstant;
+import org.springframework.security.core.context.SecurityContextHolder;
+import org.springframework.security.web.authentication.WebAuthenticationDetailsSource;
+import org.springframework.stereotype.Component;
+
+@Component
+public class UserAuthenticationFactory {
+
+    public void authenticateUser(Claims claims, HttpServletRequest request) {
+        Long userId = claims.get(AuthConstant.USER_ID_CLAIM_NAME, Long.class);
+
+        UserAuthentication authentication = UserAuthentication.createUserAuthentication(userId);
+        authentication.setDetails(new WebAuthenticationDetailsSource().buildDetails(request));
+
+        SecurityContextHolder.getContext().setAuthentication(authentication);
+    }
+}
