+++ bbangzip-auth/src/main/java/org/sopt/jwt/dto/JwtTokensDto.java
@@ -1,0 +1,10 @@
+package org.sopt.jwt.dto;
+
+import lombok.Builder;
+
+@Builder
+public record JwtTokensDto(
+        String accessToken,
+        String refreshToken
+) { }
+
