+++ bbangzip-auth/src/main/java/org/sopt/jwt/dto/ReissueJwtTokensDto.java
@@ -1,0 +1,10 @@
+package org.sopt.jwt.dto;
+
+import lombok.Builder;
+
+@Builder
+public record ReissueJwtTokensDto(
+        String accessToken,
+        String refreshToken
+) {
+}
