+++ bbangzip-auth/src/main/java/org/sopt/security/logsupport/ErrorLogger.java
@@ -1,0 +1,17 @@
+package org.sopt.security.logsupport;
+
+import lombok.extern.slf4j.Slf4j;
+import org.sopt.code.ErrorCode;
+import org.springframework.stereotype.Component;
+
+@Slf4j
+@Component
+public class ErrorLogger {
+    public void log(ErrorCode errorCode, Exception e) {
+        if (errorCode.getHttpStatus().is4xxClientError()) {
+            log.warn("Client error [{}]: {}", errorCode.getCode(), e.getMessage());
+        } else if (errorCode.getHttpStatus().is5xxServerError()) {
+            log.error("Server error [{}]: {}", errorCode.getCode(), e.getMessage(), e);
+        }
+    }
+}
