+++ bbangzip-auth/src/main/java/org/sopt/security/utils/SecurityUtils.java
@@ -1,0 +1,16 @@
+package org.sopt.security.utils;
+
+import org.sopt.exception.AuthConstant;
+import org.sopt.exception.AuthErrorCode;
+import org.sopt.exception.UnAuthorizedException;
+
+// Null check 대행할 유틸 클래스
+public class SecurityUtils {
+
+    public static Object checkPrincipal(final Object principal) {
+        if (AuthConstant.ANONYMOUS_USER.equals(principal)) {
+            throw new UnAuthorizedException(AuthErrorCode.UNAUTHORIZED);
+        }
+        return principal;
+    }
+}
